Adapt data model to include an N length effect parameter for the grass cover in/outwards, height structures and dune erosion failure mechanisms. WTI-1531

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/Create/GrassCoverErosionInwards/GrassCoverErosionInwardsFailureMechanismCreateExtensions.cs
@@ -54,7 +54,7 @@ namespace Application.Ringtoets.Storage.Create.GrassCoverErosionInwards
         {
             var metaEntity = new GrassCoverErosionInwardsFailureMechanismMetaEntity
             {
-                N = mechanism.GeneralInput.N,
+                N = (int) mechanism.GeneralInput.N,
                 DikeProfileCollectionSourcePath = mechanism.DikeProfiles.SourcePath
             };
             entity.GrassCoverErosionInwardsFailureMechanismMetaEntities.Add(metaEntity);
